fix(aws-cli): add platform guard and reliable not_if guard
Restrict Linux zip installer to deb-based distros and use Homebrew on
macOS. Replace unreliable 'which aws' shell guard with File.exist? on
the install path. Add *.zip to .gitignore to prevent stale downloads
from being tracked.

+++ cookbooks/aws-cli/default.rb
@@ -1,11 +1,16 @@
 include_recipe "dependency.rb"
 
-execute "install awscli" do
-  command <<-EOCMD
-    curl "https://awscli.amazonaws.com/awscli-exe-linux-x86_64.zip" -o "awscliv2.zip"
-    unzip awscliv2.zip
-    sudo ./aws/install
-    rm -rf ./aws
-  EOCMD
-  not_if "test -f /usr/local/bin/aws"
+case node[:platform]
+when "debian", "ubuntu", "mint", "pop"
+  execute "install awscli" do
+    command <<-EOCMD
+      curl "https://awscli.amazonaws.com/awscli-exe-linux-x86_64.zip" -o "awscliv2.zip"
+      unzip awscliv2.zip
+      sudo ./aws/install
+      rm -rf ./aws
+    EOCMD
+    not_if { File.exist?("/usr/local/bin/aws") }
+  end
+when "darwin"
+  package "awscli"
 end
